Split main into components, inject store
