--- src/gui/src/services/constants.ts
@@ -1,5 +1,0 @@
-import {InjectionKey, Ref} from 'vue'
-import {LauncherConfig} from '@/services/backend'
-
-// provide/inject typing keys
-export const CONFIG_KEY: InjectionKey<Ref<LauncherConfig>> = Symbol('config')
