+++ src/gui/src/services/index.ts
@@ -1,6 +1,6 @@
 export * as backend from './backend'
-export * as constants from './constants'
 export * as i18n from './i18n'
+export * as store from './store'
 
 export {t} from './i18n'
 export {log} from './logging'
